set pipelining to 2 as soon as we're into the main loop
this is a tiny improvement: previously the loop only started pipelining after the first command on the assumption that that was a ROWS command with a potentially long response.  but as the changed test cases demonstrate, if the key ranges were the same at both ends, then it could also just be a plain HASH command and we'd rather get on with the second HASH command that will have resulted from the initial subdivision.